Fixed the getState function in the StateServer
The getState function now returns a VIEW of the state rather than the
actual state object. This is much needed requirement because if the
function does not return a view, but returns the actual state, anyone
who has called a getState method has full access to mutate the state
BYPASSING the setState method all together. This is not at all ideal as
it fails to encapsulate the state so that it can properly be controller.
currently the state is still being exposed from the StateManager but I
am thinking of making it only possible to obtain a view from the
StateManager in future revisions.